subscriptions/serializers.py
# subscriptions/serializers.py

from rest_framework import serializers
from .models import Subscription

class SubscriptionSerializer(serializers.ModelSerializer):
    class Meta:
        model = Subscription
        fields = '__all__'

# subscriptions/views.py

from rest_framework import serializers
from .models import DeliveryLog

class DeliveryLogSerializer(serializers.ModelSerializer):
    class Meta:
        model = DeliveryLog
        fields = '__all__'
